Create

+++ Assets/MyFolder/Scripts/ScriptableObjects/Result.cs
@@ -6,12 +6,13 @@ namespace Unity1Week.ScriptableObjects
     {
         public uint KillScore;
         public TitleSettings titleSettings;
+
         public double CalcScore()
         {
-            var times = (double)10000 / (double)(titleSettings.Width * titleSettings.Height);
-            return KillScore * times * (KillScore >= titleSettings.ClearKillScore ?
-               (double)(titleSettings.LeaderCount * titleSettings.LeaderCount / 100)
-               : (double)(titleSettings.LeaderCount / 100));
+            var times = 10000 / (double) (titleSettings.Width * titleSettings.Height);
+            return KillScore * times * (KillScore >= titleSettings.ClearKillScore
+                ? titleSettings.LeaderCount * titleSettings.LeaderCount / 100
+                : (double) (titleSettings.LeaderCount / 100));
         }
     }
 }
